// src/kaspa/kaspa.d.ts
        interface UtxoProcessor {
            /**
            * @param {UtxoProcessorNotificationCallback} callback
            */
            addEventListener(callback: UtxoProcessorNotificationCallback): void;
            /**
            * @param {UtxoProcessorEventType} event
            * @param {UtxoProcessorNotificationCallback} [callback]
            */
            addEventListener<E extends keyof UtxoProcessorEventMap>(
                event: E,
                callback: UtxoProcessorNotificationCallback<E>
            )
        }


    /**
     * UtxoProcessor constructor arguments.
     * 
     * @see {@link UtxoProcessor}, {@link UtxoContext}, {@link RpcClient}, {@link NetworkId}
     * @category Wallet SDK
     */
    export interface IUtxoProcessorArgs {
        /**
         * The RPC client to use for network communication.
         */
        rpc : RpcClient;
        networkId : NetworkId | string;
    }
// ...
        export enum UtxoProcessorEventType {
            Connect = "connect",
            Disconnect = "disconnect",
            UtxoIndexNotEnabled = "utxo-index-not-enabled",
            SyncState = "sync-state",
            UtxoProcStart = "utxo-proc-start",
            UtxoProcStop = "utxo-proc-stop",
            UtxoProcError = "utxo-proc-error",
            DaaScoreChange = "daa-score-change",
            Pending = "pending",
            Reorg = "reorg",
            Stasis = "stasis",
            Maturity = "maturity",
            Discovery = "discovery",
            Balance = "balance",
            Error = "error",
        }

        /**
         * {@link UtxoProcessor} notification event data map.
         * 
         * @category Wallet API
         */
        export type UtxoProcessorEventMap = {
            "connect": IConnectEvent,
            "disconnect": IDisconnectEvent,
            "utxo-index-not-enabled": IUtxoIndexNotEnabledEvent,
            "sync-state": ISyncStateEvent,
            "server-status": IServerStatusEvent,
            "utxo-proc-start": undefined,
            "utxo-proc-stop": undefined,
            "utxo-proc-error": IUtxoProcErrorEvent,
            "daa-score-change": IDaaScoreChangeEvent,
            "pending": IPendingEvent,
            "reorg": IReorgEvent,
            "stasis": IStasisEvent,
            "maturity": IMaturityEvent,
            "discovery": IDiscoveryEvent,
            "balance": IBalanceEvent,
            "error": IErrorEvent
        }
// ...
        export type UtxoProcessorNotificationCallback<E extends keyof UtxoProcessorEventMap = keyof UtxoProcessorEventMap> = (event: UtxoProcessorEvent<E>) => void;
// ...
    export interface IUtxoContextArgs {
        /**
         * Associated UtxoProcessor.
         */
        processor: UtxoProcessor;
        /**
         * Optional id for the UtxoContext.
         * **The id must be a valid 32-byte hex string.**
         * You can use {@link sha256FromBinary} or {@link sha256FromText} to generate a valid id.
         * 
         * If not provided, a random id will be generated.
         * The IDs are deterministic, based on the order UtxoContexts are created.
         */
        id?: HexString;
    }
// ...
interface IGeneratorSettingsObject {
    /** 
     * Final transaction outputs (do not supply change transaction).
     * 
     * Typical usage: { address: "kaspa:...", amount: 1000n }
     */
    outputs: PaymentOutput | IPaymentOutput[];
    /** 
     * Address to be used for change, if any. 
     */
    changeAddress: Address | string;
    /** 
     * Priority fee in SOMPI.
     * 
     * If supplying `bigint` value, it will be interpreted as a sender-pays fee.
     * Alternatively you can supply an object with `amount` and `source` properties
     * where `source` contains the {@link FeeSource} enum.
     * 
     * **IMPORTANT:* When sending an outbound transaction (transaction that
     * contains outputs), the `priorityFee` must be set, even if it is zero.
     * However, if the transaction is missing outputs (and thus you are
     * creating a compound transaction against your change address),
     * `priorityFee` should not be set (i.e. it should be `undefined`).
     * 
     * @see {@link IFees}, {@link FeeSource}
     */
    priorityFee?: IFees | bigint;
    /**
     * UTXO entries to be used for the transaction. This can be an
     * array of UtxoEntry instances, objects matching {@link IUtxoEntry}
     * interface, or a {@link UtxoContext} instance.
     */
    entries: IUtxoEntry[] | UtxoEntryReference[] | UtxoContext;
    /**
     * Optional UTXO entries that will be consumed before those available in `entries`.
     * You can use this property to apply custom input selection logic.
     * Please note that these inputs are consumed first, then `entries` are consumed
     * to generate a desirable transaction output amount.  If transaction mass
     * overflows, these inputs will be consumed into a batch/sweep transaction
     * where the destination if the `changeAddress`.
     */
    priorityEntries?: IUtxoEntry[] | UtxoEntryReference[],
    /**
     * Optional number of signature operations in the transaction.
     */
    sigOpCount?: number;
    /**
     * Optional minimum number of signatures required for the transaction.
     */
    minimumSignatures?: number;
    /**
     * Optional data payload to be included in the transaction.
     */
    payload?: Uint8Array | HexString;

    /**
     * Optional NetworkId or network id as string (i.e. `mainnet` or `testnet-11`). Required when {@link IGeneratorSettingsObject.entries} is array
     */
    networkId?: NetworkId | string
}
// ...
export type HexString = string;
// ...
        interface RpcClient {
            /**
            * @param {RpcEventCallback} callback
            */
            addEventListener(callback:RpcEventCallback): void;
            /**
            * @param {RpcEventType} event
            * @param {RpcEventCallback} [callback]
            */
            addEventListener<M extends keyof RpcEventMap>(
                event: M,
                callback: (eventData: RpcEventMap[M]) => void
            )
        }
// ...
export type RpcEventMap = {
    "connect" : undefined,
    "disconnect" : undefined,
    "block-added" : IBlockAdded,
    "virtual-chain-changed" : IVirtualChainChanged,
    "finality-conflict" : IFinalityConflict,
    "finality-conflict-resolved" : IFinalityConflictResolved,
    "utxos-changed" : IUtxosChanged,
    "sink-blue-score-changed" : ISinkBlueScoreChanged,
    "virtual-daa-score-changed" : IVirtualDaaScoreChanged,
    "pruning-point-utxo-set-override" : IPruningPointUtxoSetOverride,
    "new-block-template" : INewBlockTemplate,
}
// ...
export type RpcEventCallback = (event: RpcEvent) => void;
// ...
export class Address {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {string} address
*/
  constructor(address: string);
/**
* @param {string} address
* @returns {boolean}
*/
  static validate(address: string): boolean;
/**
* Convert an address to a string.
* @returns {string}
*/
  toString(): string;
/**
* @param {number} n
* @returns {string}
*/
  short(n: number): string;
/**
*/
  readonly payload: string;
/**
*/
  readonly prefix: string;
/**
*/
  setPrefix: string;
/**
*/
  readonly version: string;
}
// ...
export class NetworkId {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {any} value
*/
  constructor(value: any);
/**
* @returns {string}
*/
  toString(): string;
/**
* @returns {string}
*/
  addressPrefix(): string;
/**
*/
  readonly id: string;
/**
*/
  suffix?: number;
/**
*/
  type: NetworkType;
}
// ...
export class UtxoContext {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {IUtxoContextArgs} js_value
*/
  constructor(js_value: IUtxoContextArgs);
/**
* Performs a scan of the given addresses and registers them in the context for event notifications.
* @param {(Address | string)[]} addresses
* @param {bigint | undefined} [optional_current_daa_score]
* @returns {Promise<void>}
*/
  trackAddresses(addresses: (Address | string)[], optional_current_daa_score?: bigint): Promise<void>;
/**
* Unregister a list of addresses from the context. This will stop tracking of these addresses.
* @param {(Address | string)[]} addresses
* @returns {Promise<void>}
*/
  unregisterAddresses(addresses: (Address | string)[]): Promise<void>;
/**
* Clear the UtxoContext.  Unregister all addresses and clear all UTXO entries.
* IMPORTANT: This function must be manually called when disconnecting or re-connecting to the node
* (followed by address re-registration).  
* @returns {Promise<void>}
*/
  clear(): Promise<void>;
/**
*
* Returns a range of mature UTXO entries that are currently
* managed by the UtxoContext and are available for spending.
*
* NOTE: This function is provided for informational purposes only.
* **You should not manage UTXO entries manually if they are owned by UtxoContext.**
*
* The resulting range may be less than requested if UTXO entries
* have been spent asynchronously by UtxoContext or by other means
* (i.e. UtxoContext has received notification from the network that
* UtxoEntries have been spent externally).
*
* UtxoEntries are kept in in the ascending sorted order by their amount.
* @param {number} from
* @param {number} to
* @returns {UtxoEntryReference[]}
*/
  getMatureRange(from: number, to: number): UtxoEntryReference[];
/**
* Returns pending UTXO entries that are currently managed by the UtxoContext.
* @returns {UtxoEntryReference[]}
*/
  getPending(): UtxoEntryReference[];
/**
* Current {@link Balance} of the UtxoContext.
*/
  readonly balance: Balance | undefined;
/**
* Current {@link BalanceStrings} of the UtxoContext.
*/
  readonly balanceStrings: BalanceStrings | undefined;
/**
*/
  readonly isActive: boolean;
/**
* Obtain the length of the mature UTXO entries that are currently
* managed by the UtxoContext.
*/
  readonly matureLength: number;
}
// ...
export class UtxoProcessor {
/**
** Return copy of self without private attributes.
*/
  toJSON(): Object;
/**
* Return stringified version of self.
*/
  toString(): string;
  free(): void;
/**
* @param {UtxoProcessorEventType | UtxoProcessorEventType[] | string | string[]} event
* @param {UtxoProcessorNotificationCallback | undefined} [callback]
*/
  removeEventListener(event: UtxoProcessorEventType | UtxoProcessorEventType[] | string | string[], callback?: UtxoProcessorNotificationCallback): void;
/**
* UtxoProcessor constructor.
*
*
*
* @see {@link IUtxoProcessorArgs}
* @param {IUtxoProcessorArgs} js_value
*/
  constructor(js_value: IUtxoProcessorArgs);
/**
* Starts the UtxoProcessor and begins processing UTXO and other notifications.
* @returns {Promise<void>}
*/
  start(): Promise<void>;
/**
* Stops the UtxoProcessor and ends processing UTXO and other notifications.
* @returns {Promise<void>}
*/
  stop(): Promise<void>;
/**
* @param {NetworkId | string} network_id
*/
  setNetworkId(network_id: NetworkId | string): void;
/**
*
* Set the coinbase transaction maturity period DAA score for a given network.
* This controls the DAA period after which the user transactions are considered mature
* and the wallet subsystem emits the transaction maturity event.
*
* @see {@link TransactionRecord}
* @see {@link IUtxoProcessorEvent}
*
* @category Wallet SDK
* @param {NetworkId | string} network_id
* @param {bigint} value
*/
  static setCoinbaseTransactionMaturityDAA(network_id: NetworkId | string, value: bigint): void;
/**
*
* Set the user transaction maturity period DAA score for a given network.
* This controls the DAA period after which the user transactions are considered mature
* and the wallet subsystem emits the transaction maturity event.
*
* @see {@link TransactionRecord}
* @see {@link IUtxoProcessorEvent}
*
* @category Wallet SDK
* @param {NetworkId | string} network_id
* @param {bigint} value
*/
  static setUserTransactionMaturityDAA(network_id: NetworkId | string, value: bigint): void;
/**
*/
  readonly isActive: boolean;
/**
*/
  readonly networkId: string | undefined;
/**
*/
  readonly rpc: RpcClient;
}